Add `yii\db\CacheableQueryTrait`
